pyproject.toml
#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

[project]
name = "neurowolf"
version = "2.0.0"
description = "Simple Telegram bot with Deepseek generating russian memes"
authors = [{ name = "Dmitrii Gusev", email = "amidg4x4dima@gmail.com" }]
dependencies = [
    #"ipex-llm[cpp]", # already included in intel igpu Dockerfile
    "Pillow",
    "python-telegram-bot",
    "ollama"
]

#[tool.black]
#line-length = 88
#
#[tool.isort]
#profile = "black"
#
#[tool.mypy]
#strict = true
#
#[tool.pytest.ini_options]
#testpaths = ["tests"]
#
#[tool.ruff]
#line-length = 88
